+++ source/AsioExpress/Platform/GetClockCount.hpp
@@ -7,6 +7,6 @@
 
 namespace AsioExpress {
 
-unsigned long GetClockCount(void);
+unsigned int GetClockCount(void);
 
 } // namespace AsioExpress
