+++ source/AsioExpress/Timer/StandardRepeatingTimer.hpp
@@ -26,7 +26,7 @@ public:
 
   StandardRepeatingTimer(
       boost::asio::io_service & ioService, 
-      unsigned long milliseconds) :
+      unsigned int milliseconds) :
     m_timer(ioService),
     m_milliseconds(milliseconds)
   {
@@ -41,7 +41,7 @@ public:
   }
 
   virtual void AsyncWait(
-      unsigned long milliseconds, 
+      unsigned int milliseconds, 
       CompletionHandler completionHandler)
   {
     m_timer.Start(
@@ -56,7 +56,7 @@ public:
 
 private:
   RepeatingTickCountTimer   m_timer;
-  unsigned long             m_milliseconds;
+  unsigned int             m_milliseconds;
 };
 
 } // namespace AsioExpress
